Checkpoint: cleaning up stuff
