+++ addon/operator/geopath_datastructure.py
@@ -1,12 +1,17 @@
 import bmesh
+from bmesh.types import BMFace
+import bpy
 
 from bpy_extras import view3d_utils
 from functools import reduce
 from enum import Enum
-from ..algorithms.geodesic import \
-    geodesic_walk, continue_geodesic_walk, gradient_descent
+
+from mathutils import Vector
+# from ..algorithms.geodesic_fast_marching import geodesic_walk
+from ..algorithms.geodesic_edge_flipping import geodesic_walk
 from mathutils.geometry import intersect_point_line
 from ..utility import draw
+from ..utility.geometry import create_face_with_ccw_normal
 
 
 class GeoPath(object):
@@ -16,6 +21,7 @@ class GeoPath(object):
     '''
     def __init__(self, context, selected_obj):
 
+        self.context = context
         self.selected_obj = selected_obj
         self.bme = bmesh.new()
         self.bme.from_mesh(selected_obj.data)
@@ -23,12 +29,16 @@ class GeoPath(object):
         self.bme.edges.ensure_lookup_table()
         self.bme.faces.ensure_lookup_table()
 
+        # Keep a backup of the original structure
+        self.original_bme = self.bme.copy()
+
+        self.sub_vert_undo = dict()
+
         non_tris = [f for f in self.bme.faces if len(f.verts) > 3]
         bmesh.ops.triangulate(self.bme, faces=non_tris)
 
-        self.key_points = []
+        self.key_verts = []
         self.path_segments = []
-        self.grab_undo_segment = []
 
         self.point_size = 8
         self.circle_radius = 8
@@ -38,18 +48,15 @@ class GeoPath(object):
         self.line_thickness = 3
         self.debug_color = (1, 1, 0, 1)
 
-        self.epsilon = .0000001
-        self.max_iters = 100000
         self.distance_threshold = 0.006
 
         # geos, fixed, close, far
-        self.geo_data = [None, None]
         self.hover_point_index = None
         self.selected_point_index = None
 
-        self.insert_key_point = None
+        self.insert_cursor_info = None
         self.insert_segment_index = None
-        self.is_inserting = False
+        self.insert_vert = None
 
         self.is_debugging = False
 
@@ -61,28 +68,29 @@ class GeoPath(object):
             return
 
         hit_face = self.bme.faces[face_ind]
-        self.key_points.append((hit_location, hit_face))
 
-        if len(self.key_points) < 2:
+        vert = self.decide_vert_from_face(hit_location, hit_face,
+                                          self.distance_threshold*0.5)
+
+        self.key_verts.append(vert)
+
+        if len(self.key_verts) < 2:
             return
 
         #  The elements just before the ones we just pushed
-        start_loc, start_face = self.key_points[-2]
+        start_vert = self.key_verts[-2]
 
-        geos, fixed, close, far = geodesic_walk(
-            self.bme.verts, start_face, start_loc,
-            hit_face, self.max_iters)
-
-        path_elements, path = gradient_descent(
-            geos, hit_face, hit_location, self.epsilon)
-
-        self.cleanup_path(start_loc, hit_location, path)
+        path = geodesic_walk(
+            self.bme, start_vert, vert
+        )
 
         self.path_segments.append(path)
 
-        self.geo_data.append((geos, fixed, close, far))
-
     def grab_mouse_move(self, context, x, y):
+
+        # At least one segment
+        if len(self.path_segments) == 0:
+            return
 
         hit, hit_loc, face_ind = self.raycast(context, x, y)
 
@@ -95,30 +103,42 @@ class GeoPath(object):
             self.find_keypoint_hover(hit_loc)
             return
 
-        # At least one segment
-        if len(self.path_segments) == 0:
-            return
-
         # otherwise move the selected point
         point_pos = self.selected_point_index
-        hit_face = self.bme.faces[face_ind]
+        vert_moving = self.key_verts[point_pos]
+
+        # If a vertex of the raycasted face is found in our key verts
+        # it means we might make a mess if we carry on
+        # we won't do any changes
+        if any(x != vert_moving and x in self.key_verts
+               for x in self.bme.faces[face_ind].verts):
+            # print("I won't do any grabbing, bye!")
+            return
+
+        # Before deciding, try undoing
+        if self.try_undo_subdivision(vert_moving) is True:
+            # We'll need to do another raycast to get the new hit face
+            hit, hit_loc, face_ind = self.raycast(context, x, y)
+
+        new_vert = self.decide_vert_from_face(
+            hit_loc, self.bme.faces[face_ind],
+            self.distance_threshold)
 
         # I have a segment before point
         if point_pos > 0:
-            start_loc, start_face = self.key_points[point_pos-1]
-            end_loc, end_face = self.key_points[point_pos]
+            start_vert = self.key_verts[point_pos-1]
             self.redo_geodesic_segment(
-                point_pos-1, start_loc, start_face, end_loc, end_face, 0)
+                point_pos-1, start_vert, new_vert)
 
         # I have a segment after point
-        if point_pos < len(self.key_points)-1:
-            start_loc, start_face = self.key_points[point_pos+1]
-            end_loc, end_face = self.key_points[point_pos]
+        if point_pos < len(self.key_verts)-1:
+            end_vert = self.key_verts[point_pos+1]
             self.redo_geodesic_segment(
-                point_pos, start_loc, start_face, end_loc, end_face, 1)
+                point_pos, new_vert, end_vert)
 
         # Finally move the key_point
-        self.key_points[point_pos] = (hit_loc, hit_face)
+        self.key_verts[point_pos] = new_vert
+        # print(self.key_verts)
 
     def grab_start(self):
 
@@ -128,69 +148,21 @@ class GeoPath(object):
         self.selected_point_index = self.hover_point_index
         self.hover_point_index = None
 
-        self.grab_undo_loc, self.grab_undo_face = \
-            self.key_points[self.selected_point_index]
-
-        # start, only add the segment with its same index
-        if self.selected_point_index == 0:
-            self.grab_undo_segment.append(
-                self.path_segments[self.selected_point_index]
-            )
-        # end, only the previous one
-        elif self.selected_point_index == len(self.key_points)-1:
-            self.grab_undo_segment.append(
-                self.path_segments[self.selected_point_index-1]
-            )
-        # in the middle of two segments
-        else:
-            self.grab_undo_segment.append(
-                self.path_segments[self.selected_point_index]
-            )
-            self.grab_undo_segment.append(
-                self.path_segments[self.selected_point_index-1]
-            )
-
         return True
 
     def grab_cancel(self):
 
         self.hover_point_index = None
-
-        if self.selected_point_index is None:
-            return
-
-        self.key_points[self.selected_point_index] = \
-            (self.grab_undo_loc, self.grab_undo_face)
-
-        # start, only add the segment with its same index
-        if self.selected_point_index == 0:
-            self.path_segments[self.selected_point_index] = \
-                self.grab_undo_segment[0]
-        # end, only the previous one
-        elif self.selected_point_index == len(self.key_points)-1:
-            self.path_segments[self.selected_point_index-1] = \
-                self.grab_undo_segment[0]
-        # in the middle of two segments
-        else:
-            self.path_segments[self.selected_point_index] = \
-                self.grab_undo_segment[0]
-            self.path_segments[self.selected_point_index-1] = \
-                self.grab_undo_segment[1]
-
         self.selected_point_index = None
 
         return
 
     def grab_finish(self):
-        self.grab_undo_loc = None
-        self.grab_undo_face = None
-        self.grab_undo_segment = []
 
         # Small trick to keep hovering on the point after releasing mouse
         self.hover_point_index = self.selected_point_index
 
         self.selected_point_index = None
-        self.geo_data = [None, None]
 
         return
 
@@ -224,7 +196,7 @@ class GeoPath(object):
 
         # I have a segment after point
         segment_after = None
-        if point_pos < len(self.key_points)-1:
+        if point_pos < len(self.key_verts)-1:
             segment_after = self.path_segments[point_pos]
 
         if segment_before:
@@ -233,19 +205,16 @@ class GeoPath(object):
             self.path_segments.remove(segment_after)
 
         # Remove position from keypoints
-        self.key_points.pop(point_pos)
+        self.key_verts.pop(point_pos)
 
         # Redo geodesic path if needed
         if segment_before and segment_after:
-            start_loc, start_face = self.key_points[point_pos-1]
-            end_loc, end_face = self.key_points[point_pos]
+            start_vert = self.key_verts[point_pos-1]
+            end_vert = self.key_verts[point_pos]
             # Recreate the position
             self.path_segments.insert(point_pos-1, [])
             self.redo_geodesic_segment(
-                point_pos-1, start_loc, start_face, end_loc, end_face, 0)
-
-            # Avoid having garbage in the geo cache
-            self.geo_data[0] = None
+                point_pos-1, start_vert, end_vert)
 
     def erase_cancel(self, context):
         # Reset hovering point
@@ -257,106 +226,119 @@ class GeoPath(object):
         hit, hit_loc, face_ind = self.raycast(context, x, y)
 
         if not hit:
-            self.insert_key_point = None
+            self.insert_cursor_info = None
             self.insert_segment_index = None
+            self.insert_vert = None
             context.window.cursor_set("DEFAULT")
             return
 
+        self.insert_cursor_info = (hit_loc, self.bme.faces[face_ind])
         context.window.cursor_set("NONE")
-        hit_face = self.bme.faces[face_ind]
+
+        if self.insert_vert is None:
+
+            # Try find an intersection with a segment
+            intersect_index = self.get_segment_point_intersection(
+                hit_loc, 0.0009)  # TODO: Is epsilon good enough?
+
+            # Bingo! We got an intersection (or was none)
+            self.insert_segment_index = intersect_index
+
+            return
+
+        # If we reached this point it means we're dragging
+        # the inserted point
+        if self.try_undo_subdivision(self.insert_vert) is True:
+            # We'll need to do another raycast to get the new hit face
+            hit, hit_loc, face_ind = self.raycast(context, x, y)
 
         # establish the key point
-        self.insert_key_point = (hit_loc, hit_face)
+        new_vert = self.decide_vert_from_face(
+            hit_loc, self.bme.faces[face_ind], self.distance_threshold)
 
-        if self.is_inserting:
+        # Reassign key point
+        self.key_verts[self.insert_segment_index+1] = \
+            new_vert
 
-            # Reassign key point
-            self.key_points[self.insert_segment_index+1] = \
-                self.insert_key_point
+        # First segment locations
+        start_vert = self.key_verts[self.insert_segment_index]
+        end_vert = new_vert
 
-            # First segment locations
-            start_loc, start_face = self.key_points[self.insert_segment_index]
-            end_loc, end_face = self.insert_key_point
+        # Recreate the segment
+        self.redo_geodesic_segment(
+            self.insert_segment_index,
+            start_vert, end_vert)
 
-            # Recreate the segment
-            self.redo_geodesic_segment(
-                self.insert_segment_index,
-                start_loc, start_face, end_loc, end_face, 0)
+        # Second segment locations
+        start_vert = \
+            self.key_verts[self.insert_segment_index+1]
 
-            # Second segment locations
-            start_loc, start_face = \
-                self.key_points[self.insert_segment_index+2]
+        end_vert = \
+            self.key_verts[self.insert_segment_index+2]
 
-            end_loc, end_face = \
-                self.key_points[self.insert_segment_index+1]
-
-            # Recreate the segment
-            self.redo_geodesic_segment(
-                self.insert_segment_index+1,
-                start_loc, start_face, end_loc, end_face, 1)
-
-            return
-
-        # Try find an intersection with a segment
-        intersect_index = self.get_segment_point_intersection(
-            hit_loc, 0.0009)  # TODO: Is epsilon good enough?
-
-        if (intersect_index is None):
-            self.insert_segment_index = None
-            return
-
-        self.insert_segment_index = intersect_index
+        # Recreate the segment
+        self.redo_geodesic_segment(
+            self.insert_segment_index+1,
+            start_vert, end_vert)
 
     def insert_start(self):
 
+        # Check whether when moving we reached
+        # an intersection
         if self.insert_segment_index is None:
+            return
+
+        location, face = self.insert_cursor_info
+
+        insert_vert = self.decide_vert_from_face(
+            location, face, self.distance_threshold)
+
+        # Very edge case but if we get a vertex
+        # that was already assigned this will protect
+        # from deleting and recreating geometry
+        # making the whole chain blow up
+        if insert_vert in self.key_verts:
+            # print("I wont do insertion here, bye!")
             return
 
         # Add new segment
         self.path_segments.insert(self.insert_segment_index, [])
 
         # First segment locations
-        start_loc, start_face = self.key_points[self.insert_segment_index]
-        end_loc, end_face = self.insert_key_point
+        start_vert = self.key_verts[self.insert_segment_index]
+        end_vert = insert_vert
 
         # Recreate the position
         self.redo_geodesic_segment(
             self.insert_segment_index,
-            start_loc, start_face, end_loc, end_face, 0)
+            start_vert, end_vert)
 
         # Add the new key_point
-        self.key_points.insert(self.insert_segment_index+1,
-                               self.insert_key_point)
-
-        # Avoid having garbage in the geo cache
-        self.geo_data[0] = None
+        self.key_verts.insert(self.insert_segment_index+1,
+                              insert_vert)
 
         # Second segment locations
-        start_loc, start_face = self.insert_key_point
-        end_loc, end_face = self.key_points[self.insert_segment_index+2]
+        start_vert = insert_vert
+        end_vert = self.key_verts[self.insert_segment_index+2]
 
         # Recreate the geodesic path on next segment
         self.redo_geodesic_segment(
             self.insert_segment_index+1,
-            start_loc, start_face, end_loc, end_face, 0)
+            start_vert, end_vert)
 
-        # Avoid having garbage in the geo cache
-        self.geo_data[0] = None
-
-        # Set a flag por grabbing until not released
-        self.is_inserting = True
+        # Set vertex property that will remain
+        # until we release the mouse button
+        self.insert_vert = insert_vert
 
     def insert_finish(self):
-        self.insert_key_point = None
+        self.insert_cursor_info = None
         self.insert_segment_index = None
-        self.is_inserting = False
-        self.geo_data = [None, None]
+        self.insert_vert = None
 
     def insert_cancel(self, context):
-        self.insert_key_point = None
+        self.insert_cursor_info = None
         self.insert_segment_index = None
-        self.is_inserting = False
-        self.geo_data = [None, None]
+        self.insert_vert = None
         context.window.cursor_set("DEFAULT")
 
     def toggle_debugging(self):
@@ -382,8 +364,6 @@ class GeoPath(object):
         # we can safely assume we're in the segment
         if (segment_distance <= epsilon):
             return segment_index
-            # print("Found! Segment {}, Subsegment {}".format(
-            #     segment_index, inner_segment_index))
 
         return None
 
@@ -402,58 +382,18 @@ class GeoPath(object):
                                               segment_start,
                                               segment_end)[0])).length
 
-    def redo_geodesic_segment(self, segment_pos, start_loc,
-                              start_face, end_loc, end_face,
-                              cache_pos):
+    def redo_geodesic_segment(self, segment_pos,
+                              start_vert, end_vert):
 
-        # Special case handling for weird algorithm behavior
-        should_reverse = cache_pos != 1
+        path = geodesic_walk(self.bme, start_vert, end_vert)
 
-        # Try using the cached structure before relaunching
-        # a new geodesic walk
-        cached_path = self.try_continue_geodesic_walk(
-            cache_pos, end_loc, end_face)
-
-        if cached_path:
-            self.cleanup_path(start_loc, end_loc, cached_path, should_reverse)
-            self.path_segments[segment_pos] = cached_path
-            return
-
-        geos, fixed, close, far = geodesic_walk(
-                self.bme.verts, start_face, start_loc,
-                end_face, self.max_iters)
-
-        path_elements, path = gradient_descent(
-                geos, end_face, end_loc, self.epsilon)
-
-        self.geo_data[cache_pos] = (geos, fixed, close, far)
-
-        self.cleanup_path(start_loc, end_loc, path, should_reverse)
         self.path_segments[segment_pos] = path
-
-    def try_continue_geodesic_walk(self, cache_pos, hit_loc, hit_face):
-
-        # Data was not cached
-        if self.geo_data[cache_pos] is None:
-            return None
-
-        geos, fixed, close, far = self.geo_data[cache_pos]
-
-        if not all([v in fixed for v in hit_face.verts]):
-            continue_geodesic_walk(
-                geos, fixed, close, far,
-                hit_face, self.max_iters)
-
-        path_elements, path = gradient_descent(
-            geos, hit_face, hit_loc, self.epsilon)
-
-        return path
 
     def draw(self, context, plugin_state):
 
         mx = self.selected_obj.matrix_world
 
-        points = [mx @ location for (location, face) in self.key_points]
+        points = [mx @ location.co for location in self.key_verts]
 
         point_highlight_idx = \
             self.hover_point_index if self.hover_point_index is not None \
@@ -464,7 +404,7 @@ class GeoPath(object):
                             self.point_size, self.point_color)
 
         if point_highlight_idx is not None:
-            point = self.key_points[point_highlight_idx][0]
+            point = self.key_verts[point_highlight_idx].co
             draw.draw_3d_points(context, [mx @ point],
                                 self.point_size,
                                 self.point_select_color)
@@ -473,27 +413,27 @@ class GeoPath(object):
             draw.draw_3d_circles(context, points,
                                  self.circle_radius, self.point_color)
             if point_highlight_idx is not None:
-                point = self.key_points[point_highlight_idx][0]
+                point = self.key_verts[point_highlight_idx].co
                 draw.draw_3d_circles(context, [mx @ point],
                                      self.circle_radius,
                                      self.point_select_color)
 
         elif (plugin_state == Geodesic_State.INSERT
-              and self.insert_key_point):
+              and self.insert_cursor_info):
 
             color = self.point_select_color \
                     if self.insert_segment_index is not None \
                     else self.point_color
 
-            draw.draw_3d_circles(context, [mx @ self.insert_key_point[0]],
+            draw.draw_3d_circles(context, [mx @ self.insert_cursor_info[0]],
                                  self.circle_radius, color)
-            draw.draw_3d_points(context, [mx @ self.insert_key_point[0]],
+            draw.draw_3d_points(context, [mx @ self.insert_cursor_info[0]],
                                 self.point_size, color)
 
         # Draw segments
         if len(self.path_segments):
             path = self.get_whole_path()
-            draw.draw_polyline_from_3dpoints(context, self.get_whole_path(),
+            draw.draw_polyline_from_3dpoints(context, path,
                                              self.line_color,
                                              self.line_thickness)
             # Debugging points
@@ -501,23 +441,6 @@ class GeoPath(object):
                 draw.draw_3d_points(context, path,
                                     self.point_size * .45,
                                     self.debug_color)
-
-    def cleanup_path(self, start_location, target_location,
-                     path, should_reverse=True):
-
-        # Depending on the direction of the descent
-        # we need to apply some tinkering in the direction
-        if should_reverse:
-            path.reverse()
-            path.pop(0)
-            path.pop(0)
-            path.insert(0, start_location)
-            path.append(target_location)
-        else:
-            path.pop()
-            path.pop()
-            path.insert(0, target_location)
-            path.append(start_location)
 
     def get_whole_path(self):
         mx = self.selected_obj.matrix_world
@@ -542,18 +465,140 @@ class GeoPath(object):
 
         self.hover_point_index = None
 
-        key_points = [key_point for (key_point, key_face) in self.key_points]
-
         selected_keypoints = list(
-            filter(lambda x: (x-point).length <= self.distance_threshold,
-                   key_points)
+            filter(lambda x: (x.co-point).length <= self.distance_threshold,
+                   self.key_verts)
         )
 
         if selected_keypoints:
-            self.hover_point_index = key_points.index(selected_keypoints[0])
-            # print("Point found {}".format(self.hover_point_index))
+            self.hover_point_index = \
+                self.key_verts.index(selected_keypoints[0])
+
+    def decide_vert_from_face(self, point: Vector,
+                              face: BMFace, epsilon):
+
+        bm = self.bme
+
+        # Case 1: If any of the verts in the face is close enough,
+        # start from that vert
+        distances = list(map(lambda x: (x.co-point).length,
+                         face.verts))
+        index_min = min(range(len(distances)), key=distances.__getitem__)
+
+        if distances[index_min] <= epsilon:
+            # print("Case 1: Vertex close enough")
+            return face.verts[index_min]
+
+        # Case 2: Am I close enough to an edge?
+        # Create a vertex of the collision point
+        # Create four faces out of the 2 we have connecting to the new vertex
+        # set the created vertex as seed
+        distances = list(map(lambda x: self.point_edge_distance(point, x),
+                         face.edges))
+        index_min = min(range(len(distances)), key=distances.__getitem__)
+
+        if distances[index_min] <= epsilon:
+            edge = face.edges[index_min]
+            new_vert = bm.verts.new(point)
+            faces = list(edge.link_faces)
+            for f in faces:
+                opposed_vert = [v for v in f.verts if v not in edge.verts][0]
+                bm.faces.remove(f)
+                for vert in edge.verts:
+                    create_face_with_ccw_normal(
+                        bm, new_vert, vert, opposed_vert)
+
+            # Store undo of vertex
+            self.sub_vert_undo[new_vert] = (
+                Geodesic_Subdivide.EDGE,
+                (edge.verts[0], edge.verts[1])
+            )
+
+            bm.edges.remove(edge)
+
+            self.save_bm_to_object()
+
+            # print("Case 2: Edge was close enough")
+            return new_vert
+
+        # Case 3: If not case 1 or 2
+        # Create 3 faces out of the old one, all connecting to the collision
+        new_vert = bm.verts.new(point)
+        for e in face.edges:
+            create_face_with_ccw_normal(
+                bm, e.verts[0], e.verts[1], new_vert)
+
+        bm.faces.remove(face)
+
+        self.save_bm_to_object()
+
+        # Store undo of vertex
+        self.sub_vert_undo[new_vert] = (
+            Geodesic_Subdivide.FACE,
+            (None)
+        )
+
+        # print("Case 3: Collision was quite at the center")
+        return new_vert
+
+    def save_bm_to_object(self):
+        self.bme.verts.ensure_lookup_table()
+        self.bme.edges.ensure_lookup_table()
+        self.bme.faces.ensure_lookup_table()
+
+        # Update objects back
+        current_mode = bpy.context.object.mode
+        bpy.ops.object.mode_set(mode='OBJECT')
+        self.bme.to_mesh(self.selected_obj.data)
+        self.selected_obj.data.update()
+        bpy.ops.object.mode_set(mode=current_mode)
+
+    def try_undo_subdivision(self, in_vert):
+
+        if in_vert not in self.sub_vert_undo:
+            return False  # There was not subdivision
+
+        action, args = self.sub_vert_undo[in_vert]
+
+        if action is Geodesic_Subdivide.EDGE:
+            # Verts composing the original edge
+            vert1, vert2 = args
+            verts = [e.other_vert(in_vert) for e in in_vert.link_edges]
+            faces = list(in_vert.link_faces)
+            edges = list(in_vert.link_edges)
+            for face in faces:
+                self.bme.faces.remove(face)
+            for edge in edges:
+                self.bme.edges.remove(edge)
+            for vert in [v for v in verts if v not in {vert1, vert2}]:
+                create_face_with_ccw_normal(self.bme, vert1, vert2, vert)
+
+        elif action is Geodesic_Subdivide.FACE:
+            verts = [e.other_vert(in_vert) for e in in_vert.link_edges]
+            faces = list(in_vert.link_faces)
+            edges = list(in_vert.link_edges)
+            for face in faces:
+                self.bme.faces.remove(face)
+            for edge in edges:
+                self.bme.edges.remove(edge)
+            create_face_with_ccw_normal(
+                self.bme, verts[0], verts[1], verts[2])
+
+        self.sub_vert_undo.pop(in_vert)
+        self.bme.verts.remove(in_vert)
+        self.save_bm_to_object()
+        return True
+
+    def point_edge_distance(self, point, edge):
+        return (point - (intersect_point_line(point,
+                                              edge.verts[0].co,
+                                              edge.verts[1].co)[0])
+                ).length
 
     def finish(self):
+        self.bme.free()
+        self.bme = self.original_bme
+        self.save_bm_to_object()
         self.bme.free()
 
 
@@ -562,3 +607,8 @@ class Geodesic_State(Enum):
     GRAB = 2
     ERASE = 3
     INSERT = 4
+
+
+class Geodesic_Subdivide(Enum):
+    EDGE = 1
+    FACE = 2
